backend/requirements.txt
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arm_pyart==2.0.1
attrs==25.1.0
blinker==1.9.0
boto3==1.37.8
botocore==1.37.9
Cartopy==0.24.1
certifi==2024.12.14
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmweather==0.3.2
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.10
cycler==0.12.1
dask==2025.2.0
fastapi==0.115.8
flexcache==0.3
flexparser==0.4
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
h5netcdf==1.5.0
h5py==3.13.0
httpcore==1.0.7
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
lat_lon_parser==1.3.1
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mda-xdrlib==0.2.0
MetPy==1.6.3
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
netCDF4==1.7.2
nexradaws==1.1
numpy==2.2.3
open-radar-data==0.4.3
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.1.0
Pint==0.24.4
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
propcache==0.3.0
pycairo==1.27.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
pyparsing==3.2.1
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
s3fs==2025.2.0
s3transfer==0.11.4
scipy==1.15.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
toolz==1.0.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
xarray==2025.1.2
xmltodict==0.14.2
xradar==0.9.0
yarl==1.18.3
